Move models to Models directory and refactor imports
