+++ app/Http/Controllers/RepositoryController.php
@@ -2,9 +2,9 @@
 
 namespace App\Http\Controllers;
 
-use App\Project;
-use App\Repository;
-use App\Suite;
+use App\Models\Project;
+use App\Models\Repository;
+use App\Models\Suite;
 use Illuminate\Http\Request;
 
 class RepositoryController extends Controller
