+++ app/Http/Controllers/TestCaseController.php
@@ -2,10 +2,10 @@
 
 namespace App\Http\Controllers;
 
-use App\Project;
-use App\Repository;
-use App\Suite;
-use App\TestCase;
+use App\Models\Project;
+use App\Models\Repository;
+use App\Models\Suite;
+use App\Models\TestCase;
 use Illuminate\Http\Request;
 
 class TestCaseController extends Controller
