+++ app/Http/Controllers/TestRunController.php
@@ -2,12 +2,12 @@
 
 namespace App\Http\Controllers;
 
-use App\Project;
-use App\Repository;
-use App\Suite;
-use App\TestCase;
-use App\TestPlan;
-use App\TestRun;
+use App\Models\Project;
+use App\Models\Repository;
+use App\Models\Suite;
+use App\Models\TestCase;
+use App\Models\TestPlan;
+use App\Models\TestRun;
 use Illuminate\Http\Request;
 
 class TestRunController extends Controller
