+++ resources/views/repository/edit_page.blade.php
@@ -1,4 +1,4 @@
-@php use App\Repository;use Illuminate\Support\MessageBag;
+@php use App\Models\Repository;use Illuminate\Support\MessageBag;
 /**
  * @var Repository $repository
  * @var MessageBag $errors
