+++ resources/views/repository/test_cases_list.blade.php
@@ -1,4 +1,4 @@
-@php use App\Enums\CasePriority;use App\Repository;use App\TestCase;
+@php use App\Enums\CasePriority;use App\Models\Repository;use App\Models\TestCase;
 /**
  * @var TestCase[] $testCases
  * @var Repository $repository
